% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.elicitEdgeProb}
\alias{print.elicitEdgeProb}
\title{Print Method for elicitEdgeProb Objects}
\usage{
\method{print}{elicitEdgeProb}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"elicitEdgeProb"}.}

\item{...}{Further arguments passed to or from other methods (unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Compact display of the elicited prior edge-inclusion probabilities.
}
