% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.elicitEdgeProbLite}
\alias{print.elicitEdgeProbLite}
\title{Print Method for elicitEdgeProbLite Objects}
\usage{
\method{print}{elicitEdgeProbLite}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"elicitEdgeProbLite"}.}

\item{...}{Further arguments passed to or from other methods (unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Compact display of the elicited prior edge-inclusion probabilities.
}
